Internal 184 log out redirect (#198)
* converted useInterval to timeMod, extracted dashboard styles, fixed useEffect bug of excessive calling

* broke styles into modules, formatting and addition of card info when not logged in

* removed logic to check if local for login, added else statement for rerender to display login message in uploadplugin and sidebar components

* fixed duplicate use state bug

* fixed grapl header styling for engagementview

* refactoring, fixed key errors for tables

* changed color of button

* black formatting
