+++ src/js/engagement_view/src/components/makeStyles/GraphVizStyles.tsx
@@ -1,0 +1,92 @@
+import {createStyles, makeStyles, Theme} from "@material-ui/core/styles";
+
+const drawerWidth = 500;
+
+export const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+        root: {
+            display: "flex"
+        },
+        button: {
+            backgroundColor: "#42C6FF",
+            margin: "0.25rem",
+            padding: "0.20rem",
+        }, 
+        loggedIn: {
+            display: "flex",
+            alignItems: "flex-end",
+            justifyContent: "flex-end",
+        },
+        appBar: {
+            transition: theme.transitions.create(["margin", "width"], {
+            easing: theme.transitions.easing.sharp,
+            duration: theme.transitions.duration.leavingScreen
+            })
+        },
+        appBarShift: {
+            width: `calc(100% - ${drawerWidth}px)`,
+            marginLeft: drawerWidth,
+            transition: theme.transitions.create(["margin", "width"], {
+            easing: theme.transitions.easing.easeOut,
+            duration: theme.transitions.duration.enteringScreen
+            })
+        },
+        menuButton: {
+            marginRight: theme.spacing(2),
+            color: "#42C6FF"
+        },
+        hide: {
+            display: "none"
+        },
+        drawer: {
+            width: drawerWidth,
+            flexShrink: 0
+        },
+        drawerPaper: {
+            width: drawerWidth
+        },
+        drawerHeader: {
+            display: "flex",
+            alignItems: "center",
+            padding: theme.spacing(0, 1),
+            // necessary for content to be below app bar
+            ...theme.mixins.toolbar,
+            justifyContent: "flex-end"
+        },
+        content: {
+            flexGrow: 1,
+            padding: theme.spacing(3),
+            transition: theme.transitions.create("margin", {
+            easing: theme.transitions.easing.sharp,
+            duration: theme.transitions.duration.leavingScreen
+            }),
+            marginLeft: -drawerWidth
+        },
+        contentShift: {
+            transition: theme.transitions.create("margin", {
+            easing: theme.transitions.easing.easeOut,
+            duration: theme.transitions.duration.enteringScreen
+            }),
+            marginLeft: 0,
+        },
+        lensName:{
+            color:"#EAFDFF",
+            fontSize: "1.5rem", 
+        },
+        header:{
+            fontSize: "35px",
+        }, 
+        headerContent: {
+            width: "100vw",
+            display: "flex",  
+            justifyContent: "space-between",
+        },
+        close:{
+            color:"#42C6FF",
+        },
+        link: {
+            color: "#42C6FF", 
+            textDecoration: "none"
+        }
+    })
+);
