+++ src/js/engagement_view/src/components/makeStyles/NodeTableStyles.tsx
@@ -1,0 +1,17 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles({
+    root: {
+        fontSize: ".75em"
+    },
+    nodeTable: {
+        minWidth: 450,
+    }, 
+    tableHeader: {
+        fontSize: ".85rem", 
+        backgroundColor: "#595959"
+    }, 
+    nodeTableData: {
+        fontSize: ".75rem"
+    },
+});
