+++ src/js/engagement_view/src/components/reusableComponents/GraplHeader.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import AppBar from "@material-ui/core/AppBar";
 import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
-import {createStyles, makeStyles, Theme} from "@material-ui/core/styles";
+import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
 import Home from '@material-ui/icons/Home';
-import {Link} from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 const useStyles = makeStyles(
     (theme: Theme) =>
@@ -22,7 +22,6 @@ const useStyles = makeStyles(
         }
     )
 );
-
 
 type GraplHeaderProps = {
     displayBtn: boolean
