+++ src/js/graphql_endpoint/modules/schema.js
@@ -420,6 +420,7 @@ const handleLensScope = async (parent, args) => {
     const lens_name = args.lens_name;
 
     const lens = await getLensByName(dg_client, lens_name);
+
     lens["scope"] = lens["scope"] || [];
 
     for (const node of lens["scope"]) {
