+++ src/python/grapl-model-plugin-deployer/src/grapl_model_plugin_deployer.py
@@ -323,10 +323,9 @@ def requires_auth(path):
             if app.current_request.method == "OPTIONS":
                 return respond(None, {})
 
-            if not IS_LOCAL:  # For now, disable authentication locally
-                if not check_jwt(app.current_request.headers):
-                    LOGGER.warning("not logged in")
-                    return respond("Must log in")
+            if not check_jwt(app.current_request.headers):
+                LOGGER.warning("not logged in")
+                return respond("Must log in")
             try:
                 return route_fn()
             except Exception as e:
